fix(thorvg): compatibility with IDF v5.4-dev
IDF has started detecting sections generated when compiling with -fPIC
and reports that as an error.

We were already passing b_staticpic=false to Meson, however turns out
this had no effect, because Meson always turns on fPIC when building
for android platform, and we have been setting platform to "android"
in our cross_file.txt.

After digging a bit, it seems that the choice of "android" platform
was a hack to bypass pthread library detection which Thorvg does for
other platforms, but skips for android and windows.

This commit changes cross_file.txt to contain the absolute path of the
directory where libpthread.a is built, so now thorvg can correctly
detect pthread, and there is no need to pretend that we are building
for android anymore.

This commit also refactors the way cross_file.txt is generated:
- started using a template in cross_file.txt.in, to be able to expand
  generator expressions using file(GENERATE)
- replaced manual list concatenation with list(JOIN)
- removed "replace_tool_name" since all the absolute tool paths are
  already known