ci: fix pyclang install
The issue is that recent modifications[1] to the ESP-IDF export scripts
allow them to identify the export script path only for bash and zsh. For
other shells, the export must be executed from within the ESP-IDF
directory.

The problem is because the default shell in a GitHub step inside a
container is sh instead of bash[2].

> The default shell for run steps inside a container is sh instead of
bash. This can be overridden with jobs.<job_id>.defaults.run or
jobs.<job_id>.steps[*].shell.

[1] https://github.com/espressif/esp-idf/commit/cad15b50c40ba1fcb1f63c106805f498c3fc971c
[2] https://docs.github.com/en/actions/writing-workflows/workflow-syntax-for-github-actions#jobsjob_idcontainer

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>